Simplify imports with es6 syntax (#556)
* Simplify imports with es6 syntax

* Fix env loading (#557)

* Fix environment config loading

Changes were made recently to take advantage of new syntax features in
ES6.  Since ES6 module imports are hoisted, they are run before any
code. This causes conflicts with the load order of our dotenv system,
which made it so the `config/custom-environment-variables.json` did not
get loaded in time.

* Add comment on importance of import order