Properly populate transaction fees (#1152)
Populate transaction fees

This commit introduces a data change to the transactions table. From now one, all the fields that are not zero are going to be present in both `credit` and `debit` transactions and fees are always negative for both the `User Ledger` and the `Collective Ledger`. Fees will only be positive for `Payment Method Ledgers` that are not being introduced right now but are certainly planned for the future.

Each row changed in the database will have its data field updated with which fee field was updated and it will contain its old value and the new value associated with it. To find out which rows were updated, use the following SQL:

  SELECT count(*) FROM public."Transactions"
  WHERE data->>'migration' is not null;