Script for setting up development database
The script `db_setup.js` is responsible for the following operations:

 * Create user (`config/<env>.json:database/username`)
 * Create database (`config/<env>.json:database/database`)
 * Create `POSTGIS` extension in the newly created database
 * Load development database into the newly created database

It's written in JavaScript mainly to share the configuration system
with the application and allow the environment setup scripts to
support different values for different environments.

This script requires a new key in the configuration file of the
environment. The `maintenancedb` key takes the same parameters as the
already existing `database` key. But it's used to execute parameters
on behalf of the maintenance user, like creating and deleting
databases, etc. E.g.:

  "maintenancedb": {
    "database": "postgres",
    "username": "postgres",
    "password": "",
    "options": { "host": "postgres" }
  },