Fix customizing db args on db:migrate:dev & db:migrate:dev:undo
The argument -l (local) was added to the sequelize.sh call in
package.json for the commands db:migrate:dev, db:migrate:dev:undo,
db:migration:create and db:seed (not sure if it's working either way).

The command db:clean (runs on postdb:migrate:dev) needed the variables
to be forwarded in their declaration in the file package.json.