fix(expenses,Collective.addHost): promises that should return
With the refactoring on the previous commit I found some promises that
touch the database returning too early and now they're being properly
`await`ed.

The notification system still doesn't wait for external systems. The
promises that were changed on this commit are related to creating the
entry on the `Activity` table. I bet these aren't causing any problems
on prod but they make the test suite very brittle.

The one promise not being waited for found on the `Collective.addHost`
method looked a little more serious and this might actually fix a real
production bug.