Move PaymentMethodLabel to a lib function and add it to OrderForm (i18n)
The fact that we use these labels inside `<select/>` options prevent us
from implementing this as a React component as for now React does not
support having components inside `<option/>` tags, even if the component
returns only strings.

[This message](https://github.com/facebook/react/issues/13586#issuecomment-419490956)
explains why its not supported (though it has been in the past) and why
it may not be in a near future.