Unify the way collective transactions are loaded to include virtualcards
As of today `collective-report`, model and graphql api each have they own way of loading collective transactions. But now we have virtual cards, there's a little business logic we need to add to ensure we load the proper transactions for collective.

When fetching collective transactions with the new function, we will now include:
- Debit transactions made by this collective without using a virtual card
- Debit transactions made using a virtual card from this collective
- Credit transactions made to this collective