fix: clean error message when changing step in createOrderNewFlow
The idea is being consistent, it is simpler to understand that the 
message error is cleaned up every time we change between steps than 
adding a new conditional for fixing only the card error.

This aims to fix the UX problem, it simply doesn't makes sense to show 
validation errors from different step.

We could also setState everytime but that would cause rendering, I'm 
adding a small conditional to avoid that.

Fixes: https://github.com/opencollective/opencollective/issues/1698