fix(ContributionFlow): Properly create organization
Organization website was not required in the component, but was required in the flow - resulting in organization not being properly created when going to the next step.

Also errors were not properly handled and it was possible to go to the next step if organization creation failed.

Fix https://github.com/opencollective/opencollective/issues/1643