fix(src/graphql/mutations.js src/graphql/queries.js): make pledge appear after submitting for collec
Update mutation for creation of order (pledge) to include changes to the collective information in
Apollo cache that is newly out of sync with GraphQL server, to include new pledge to show in the UI.
This is done via the update function provided by the Mutation API
https://www.apollographql.com/docs/react/essentials/mutations#update. Also, added a new query to
retrieve the collective currently in cache.

BREAKING CHANGE: For an unknown reason, the Collective data seems to be stripped away during
re-hydration. Similar to https://github.com/opencollective/opencollective/issues/1872