chore(StyledModal): API improvements
StyledModal relies on `Container`s, so by just passing the props to the wrapper we can set a lot more useful properties, like `maxWidth` and `maxHeight`.

This is important because `width` and `height` are not enough to make something responsive: we can't expect the height of a modal to be the same on mobile and desktop.