fix(FAQ): Use FAQ as title instead of FAQ's
"FAQ" is the most common spelling for a single list of questions, as the Q
already means "Questions", so it is already plural and does not need an
apostrophe. "FAQs" is best used for multiple _lists_ of questions (not a single
list that happens to have multiple questions), which is the case in a few spots
here. See the following resources for further evidence:

* https://en.wikipedia.org/wiki/FAQ
* https://english.stackexchange.com/questions/503/what-is-the-correct-way-to-pluralize-an-acronym-initialism