Dashboard menu items re-organization (#9459)
For https://github.com/opencollective/opencollective/issues/7029

# Description

Updates the menu in Dashboard.

Also changes the Dashboard Sections to rely less on a common 'account' from the dasbhoard query and instead use the `accountSlug`, this to move down the data fetching to the pages/sections themselves to make them less error-prone in terms of mismatch between expected data and what is sent in, and prepare for moving the sections into individual pages.

# Screenshots

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/zQq4TRpYwmRLp4NBVpwm/96da3784-0d62-4955-9160-114e57d505f2/image.png)

![Screenshot 2023-10-16 at 14.33.03.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/zQq4TRpYwmRLp4NBVpwm/99115acf-00d7-423e-bc1e-568bbfa31bed/Screenshot%202023-10-16%20at%2014.33.03.png)