Fix 2fa input on mobile (#9900)
Fix for the 2fa Authenticator code input allow the underlying input to be focusable (to allow pasting 2fa codes on mobile). 

The input was set as "opacity-0" to hide the blue background on the input set by the 1password browser extension that hid the individual code characters as the input was rendered above.

This fix changes the z-index value of the code characters instead, and setting "pointer-events-none" on them instead, to have the underlying input to receive the pointer events instead.