chore(menu-items.tsx): update href values to '#' to disable links temporarily
The href values for the Blog and About links in the landingPageItems array and the Use cases link in the regularFooterItems array have been updated to '#' to disable the links temporarily. This is done to prevent users from navigating to external pages that are not yet available or under construction.