fix: use same fonts in all SVG buttons
Currently, donate-button-blue.svg and donate-button-white.svg fall back
to Arial but contribute-button-blue.svg and contribute-button-white.svg
don't. As a result, on systems that don't have the first two fonts
installed, the buttons use different fonts and look different. This
seems to be the case on macOS.

Now, all buttons should be using the same fonts.